requirements.txt
Django>=4.2
eth-account>=0.8.0
google-genai
requests